chore: migrate three tests to jest
