+++ jest-tests/add-action-skip-function/add-action-skip-function.spec.js
